nat: fix the wrong unformat type
The unformat type for "%d" should be u32 or int.

Type: fix

+++ src/plugins/nat/nat44-ed/nat44_ed_cli.c
@@ -180,7 +180,7 @@ snat_set_log_level_command_fn (vlib_main_t * vm,
 {
   unformat_input_t _line_input, *line_input = &_line_input;
   snat_main_t *sm = &snat_main;
-  u8 log_level = NAT_LOG_NONE;
+  u32 log_level = NAT_LOG_NONE;
   clib_error_t *error = 0;
 
   if (!unformat_user (input, unformat_line_input, line_input))
